Large number of changes
Ton of crap changed.  Abstracted the enumeration of the weapons into
concrete weapons.  Added a new weapon type fist.

Added the concept of a gameboard with X & Y coordinates (probably will
be renamed to room) .  All objects that will be drawn on the board are
now gameboard objects so the controller can polymorphically iterate
through an array ( or list ) of them...

Added Use and Look functionality to all