Ton  of changes
Updated GameWindow so that it can be called as an object to another
class so that the draw methods are available to all classes of the game.

Updated TakeTurn in Controller to allow for each character on the board
to take a turn (Player character, then Mobs).

Added an inventory to PlayerCharacter.

Made it so a new PlayerCharacter comes into the game with one randomly
generated weapon and his/her fists. (Rad!)

Implemented (most of) PlayerCharacter's takeTurn.